+++ src/components/misc/to-top-button.tsx
@@ -2,14 +2,16 @@ import { faChevronUp } from '@fortawesome/free-solid-svg-icons/faChevronUp'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import LayoutStyles from '@stratego/styles/modules/Layout.module.sass'
 import { motion } from 'framer-motion'
+import dynamic from 'next/dynamic'
 import PropTypes from 'prop-types'
 import { useMemo, useRef, type FC } from 'react'
-import Button from 'react-bootstrap/Button'
 import { useWindowScroll } from 'react-use'
 
 type ToTopButtonProps = {
   show?: boolean
 }
+
+const Button = dynamic(() => import('react-bootstrap/Button'))
 
 const ToTopButton: FC<ToTopButtonProps> = ({ show = true }) => {
   const { y: scrollPosition } = useWindowScroll()
